Reorganise files and add setup.py
This makes the script segy2segy.py executable on the command line.
